feat : 그림일기 작성 API 구현 / Challenge Response 구조 변경

+++ src/main/java/com/hana/api/account/dto/response/AccountLogResponse.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 @Setter
 public class AccountLogResponse {
     private String historyOpposit;
-    private Long historyAmount;
+    private Integer historyAmount;
     private Integer historyBusinessCode;
     private Long historyBeforeBalance;
     private Long historyAfterBalance;
